Fix external node enum conversion

+++ Sources/EffeTuneLive/DSP/Designers/BandFIRPEQDesigner.swift
@@ -919,7 +919,7 @@ final class BandFIRPEQDesigner: ObservableObject {
                        outputBus: node.outputBus,
                        channelSpec: node.channelSpec,
                        sectionGate: node.sectionGate,
-                       kind: ET_PIPE_NODE_NATIVE,
+                       kind: UInt8(ET_PIPE_NODE_NATIVE),
                        externalIndex: 0)
         }
         nodes.withUnsafeBufferPointer { ETPipeline_Publish($0.baseAddress, UInt32($0.count)) }
